+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 most recent version is listed first.
 
 
+## **version:** v0.1.6
+bugfix, make sure QueueUrl is always available: https://github.com/komuw/wijisqs/pull/23
+
 ## **version:** v0.1.5
 - bugfix; fix keyError on empty queues : https://github.com/komuw/wijisqs/pull/18
 - for the same instance of a `wijisqs.SqsBroker`; create queue and tag queue should only run once : https://github.com/komuw/wijisqs/pull/19